refactor: entity -> domain 패키지명 변경
