+++ src/main/kotlin/com/example/kbocombo/domain/PlayerDetailPosition.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.entity
+package com.example.kbocombo.domain
 
 enum class PlayerDetailPosition {
     PITCHER,
