+++ src/main/kotlin/com/example/kbocombo/domain/Team.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.entity
+package com.example.kbocombo.domain
 
 enum class Team {
     NC,
